fix: remove error thrown in useMessageContext
This error is intended to give users a hint when using a context outside
of its provider.

However, the MessageContext specifically is something we use ourselves
in the Reply component, as well as multiple child components of it, both
as part of the MessageList (within Message context) and in the
MessageInput component (outside Message context).

I've considered two approaches:

1. Removing the error - the idea being that the probable value of it is
low enough in this context for it not to catch any possible bugs for any
users
2. Creating a secondary `useMessageContextDangerously` or something
along those lines for use in the specific case of the MessageInput
component.

Alternative 2 works, but seems unneccessary/redundant. However, I am
open for input/other ideas on this.